Replace custom code with QSvgGenerator for SVG.
Previously, Dunnart would attempt to save to a mixed SVG format with
it's own extra nodes and attributes.  This meant that most canvas
objects needed two sets of drawing code: one for the canvas and one to
output themselves to SVG.  This change makes use of QSvgGenerator to
output the SVG using the canvas drawing code.  This remove a lot of
needless complication and paves the way to generalise and clean up the
loading and saving code to use plugins and allow more file formats.