Don't constrain directed cycles during flow layout.
During flow layout, directed edges are constrained to point downwards.
This would cause unsatisfiable constraints if there are cycles.  Thus,
we detect strongly connected components and don't create downward
constraints for a connector if its two endpoint nodes are within the
same strongly connected component.  In this case, none of the connectors
in the cycle will have downward edge constraints applied to them.